+++ src/components/SearchBar.js
@@ -1,4 +1,4 @@
-import "./SearchBar.css";
+import "../styles/SearchBar.css";
 
 function SearchBar() {
   return (
